refactor: mcp_server module becomes a package (#286)
The 2,100-line module carried nine responsibilities; it is now a
package split along them, behavior-preserving:

- schemas.py: the _TOOLS declarations, _TOOL_SCHEMAS, compiled
  validators, _USER_COMMON_PROPERTIES - schema content untouched.
- normalize.py: every argument pre-validation helper plus the
  update_settings field/normalizer tables (their parity contract with
  tests/test_settings_plumbing_parity.py is unchanged).
- serializers.py: _user_to_dict, _client_to_dict.
- handlers_users/clients/tokens/config.py: the 25 _tool_* handlers by
  domain, bodies unchanged (including the #279 boundary and #229
  revision-precondition comments).
- __init__.py: the two MUTABLE process globals (_config,
  _readonly_mode - tests reset them by attribute assignment on
  nanoidp.mcp_server, which still works because the package module IS
  that attribute namespace), the guards, _reject/call_tool dispatch,
  the Server construction, main(), and re-imports keeping every
  pre-split import path alive.
- __main__.py: python -m nanoidp.mcp_server for a package (the smoke
  test's invocation; the nanoidp-mcp console script is untouched).

verify_secret moved to a new framework-free nanoidp.security
(routes/_auth.py re-exports it): the stdio MCP process no longer
imports Flask at all, pinned by tests/test_mcp_no_flask.py in a clean
subprocess. The #283 call-site registry follows the move
(mcp_server/handlers_tokens.py::_tool_generate_token).

1874 tests, mypy/ruff/import-linter clean, MCP stdio smoke green.
